Make img processing optional #1044
