+++ server/src/handlers/download.rs
@@ -2,10 +2,12 @@ use crate::{appstate::AppState, errors::AtomicServerResult, helpers::get_client_
 use actix_files::NamedFile;
 use actix_web::{web, HttpRequest, HttpResponse};
 use atomic_lib::{urls, Resource, Storelike};
-use image::GenericImageView;
-use image::{codecs::avif::AvifEncoder, ImageReader};
+
 use serde::Deserialize;
 use std::{collections::HashSet, io::Write, path::PathBuf};
+
+#[cfg(feature = "image")]
+use crate::handlers::image::{is_image, process_image};
 
 #[serde_with::serde_as]
 #[serde_with::skip_serializing_none]
@@ -71,11 +73,14 @@ pub fn download_file_handler_partial(
         return Ok(file.into_response(req));
     }
 
-    if !is_image(&file_path) {
-        return Err("Quality or with parameter are not supported for non image files".into());
+    // only if image feature flag is on
+    #[cfg(feature = "image")]
+    {
+        if !is_image(&file_path) {
+            return Err("Quality or with parameter are not supported for non image files".into());
+        }
+        process_image(&file_path, &processed_file_path, params)?;
     }
-
-    process_image(&file_path, &processed_file_path, params)?;
 
     let file = NamedFile::open(processed_file_path)?;
     Ok(file.into_response(req))
@@ -108,57 +113,6 @@ pub fn build_prossesed_file_path(
     processed_file_path.set_extension(format);
 
     Ok(processed_file_path)
-}
-
-fn is_image(file_path: &PathBuf) -> bool {
-    if let Ok(img) = image::open(file_path) {
-        return img.dimensions() > (0, 0);
-    }
-    false
-}
-
-fn process_image(
-    file_path: &PathBuf,
-    new_path: &PathBuf,
-    params: &DownloadParams,
-) -> AtomicServerResult<()> {
-    let format = get_format(params)?;
-    let quality = params.q.unwrap_or(100.0).clamp(0.0, 100.0);
-
-    let mut img = ImageReader::open(file_path)?
-        .with_guessed_format()?
-        .decode()
-        .map_err(|e| format!("Failed to decode image: {}", e))?;
-
-    if let Some(width) = &params.w {
-        if *width < img.dimensions().0 {
-            img = img.resize(*width, 10000, image::imageops::FilterType::Lanczos3);
-        }
-    }
-
-    if format == "webp" {
-        let encoder = webp::Encoder::from_image(&img)?;
-        let webp_image = match params.q {
-            Some(quality) => encoder.encode(quality),
-            None => encoder.encode(75.0),
-        };
-
-        let mut file = std::fs::File::create(new_path)?;
-        file.write_all(&webp_image)?;
-
-        return Ok(());
-    }
-
-    if format == "avif" {
-        let mut file = std::fs::File::create(new_path)?;
-        let encoder = AvifEncoder::new_with_speed_quality(&mut file, 8, quality as u8);
-        img.write_with_encoder(encoder)
-            .map_err(|e| format!("Failed to encode image: {}", e))?;
-
-        return Ok(());
-    }
-
-    Err(format!("Unsupported format: {}", format).into())
 }
 
 fn create_processed_folder_if_not_exists(base_path: &PathBuf) -> AtomicServerResult<()> {
